+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Adjusted the ratelimit so that it works better with the actual Discord rate limit.
 - Ensured that a user who is spamming commands will not be ratelimited multiple times for a single "spam offence".
 - Premium subscribers now gain access to unlimited role rewards.
+- $addpoints command has been created. Owner only.
 
 ### Version 2.7.3
 - Fixed a bug where commands with reactions attached to them would not execute.